+++ tests/split/b-chunk.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # test splitting into 3 chunks
 
-# Copyright (C) 2010-2013 Free Software Foundation, Inc.
+# Copyright (C) 2010-2016 Free Software Foundation, Inc.
 
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -23,28 +23,44 @@ print_ver_ split
 # in which case no data is extracted, or empty files are written
 split -n 10 /dev/null || fail=1
 test "$(stat -c %s x* | uniq -c | sed 's/^ *//; s/ /x/')" = "10x0" || fail=1
-rm x??
+rm -f x??
+
+# When extracting K of N where N > file size
+# no data is extracted, and no files are written
+split -n 2/3 /dev/null || fail=1
+returns_ 1 stat x?? 2>/dev/null || fail=1
 
 # Ensure --elide-empty-files is honored
 split -e -n 10 /dev/null || fail=1
-stat x?? 2>/dev/null && fail=1
+returns_ 1 stat x?? 2>/dev/null || fail=1
 
-printf '1\n2\n3\n4\n5\n' > in || framework_failure_
+printf '1\n2\n3\n4\n5\n' > input || framework_failure_
 
-split -n 3 in > out || fail=1
-split -n 1/3 in > b1 || fail=1
-split -n 2/3 in > b2 || fail=1
-split -n 3/3 in > b3 || fail=1
-printf '1\n2' > exp-1
-printf '\n3\n' > exp-2
-printf '4\n5\n' > exp-3
+for file in input /proc/version /sys/kernel/profiling; do
+  test -f $file || continue
 
-compare exp-1 xaa || fail=1
-compare exp-2 xab || fail=1
-compare exp-3 xac || fail=1
-compare exp-1 b1 || fail=1
-compare exp-2 b2 || fail=1
-compare exp-3 b3 || fail=1
-test -f xad && fail=1
+  split -n 3 $file > out || fail=1
+  split -n 1/3 $file > b1 || fail=1
+  split -n 2/3 $file > b2 || fail=1
+  split -n 3/3 $file > b3 || fail=1
+
+  case $file in
+    input)
+      printf '1\n2' > exp-1
+      printf '\n3\n' > exp-2
+      printf '4\n5\n' > exp-3
+
+      compare exp-1 xaa || fail=1
+      compare exp-2 xab || fail=1
+      compare exp-3 xac || fail=1
+      ;;
+  esac
+
+  compare xaa b1 || fail=1
+  compare xab b2 || fail=1
+  compare xac b3 || fail=1
+  cat xaa xab xac | compare - $file || fail=1
+  test -f xad && fail=1
+done
 
 Exit $fail
